requirements.txt
aiobotocore==1.4.2
aiohttp==3.8.0
aioitertools==0.8.0
aiosignal==1.2.0
APScheduler==3.8.1
arabic-reshaper==2.1.3
asgiref==3.4.1
async-generator==1.10
async-timeout==4.0.0
attrs==21.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
botocore==1.20.106
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2.8
django-crispy-forms==1.13.0
django-formtools==2.3
django-ipware==4.0.0
django-model-utils==4.2.0
django-notifications-hq==1.6.0
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-session-timeout==0.1.0
frozenlist==1.2.0
fsspec==2021.10.1
future==0.18.2
h11==0.12.0
html5lib==1.1
idna==3.3
jmespath==0.10.0
jsonfield==3.1.0
lxml==4.6.4
multidict==5.2.0
mysql-connector-python==8.0.27
mysqlclient==2.0.3
oauthlib==3.1.1
openpy==0.0.2
outcome==1.1.0
parser-libraries==3.7
phonenumbers==8.12.36
Pillow==8.4.0
protobuf==3.19.0
pycparser==2.20
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
PyPDF2==1.26.0
pyTelegramBotAPI==4.1.1
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.2
requests==2.26.0
requests-oauthlib==1.3.0
s3fs==2021.10.1
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
swapper==1.1.2.post1
trio==0.19.0
trio-websocket==0.9.2
twilio==7.2.0
typing-extensions==3.10.0.2
tzdata==2021.5
tzlocal==4.0.2
urllib3==1.26.7
webencodings==0.5.1
wget==3.2
wrapt==1.13.3
wsproto==1.0.0
xhtml2pdf==0.2.5
yarl==1.7.2
